Brave lads! for we have won the battle. The bug army has had its first wave destroyed. rest quickly and arm thyself once more. For the first light of morning is upon us and we have yet to fight the toughest foes...... So now all the new code compiles.... and is most likely wrong!!!!!!!! WOOOOOOOOOOOOOOOOOOOOOOOHOOOOOOOOOOOOOOOOOOOOOO

+++ src/georouting/vanet/voronoi/VoronoiEvent.java
@@ -8,6 +8,6 @@ public abstract class VoronoiEvent implements Comparable<VoronoiEvent>
 
     public int compareTo(VoronoiEvent e)
     {
-        return _y - e._y;
+        return (int)(_y - e._y);
     }
 }
